Add very basic JACK transport support
This is just enough to connect Synfig to a running JACK transport and
control it from the buttons in the canvas view. It only works one way,
Synfig controls the JACK transport.

Also add jackaudio as a build input for the nix expression, to make it
easy for end users and developers to test it out.